requirements.txt
joblib==1.2.0
matplotlib==3.7.2
numpy==1.23.5
pandas==2.0.3
PyGithub==2.1.1
python_dateutil==2.8.2
PyYAML==6.0.1
Requests==2.31.0
sqlitedict==2.1.0
urllib3==1.26.16
